scraper/pyproject.toml
[project]
name = "stock-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "playwright>=1.54.0",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.8",
]
